+++ R/check_plan.R
@@ -53,7 +53,7 @@ check_plan <- function(tests = test_db(), defaults = list(), timeout = getOption
     
     ## All combinations of arguments to test over
     if (length(test$args) == 0) {
-     sets_of_args <- as.data.frame(defaults)
+     sets_of_args <- as.data.frame(defaults, stringsAsFactors = FALSE)
     } else {
      sets_of_args <- do.call(expand.grid, test$args)
     }