+++ inst/test-db/resolve.R
@@ -1,16 +1,16 @@
-make_test(title = "resolve()", args = list(lazy = c(FALSE, TRUE), value = c(FALSE, TRUE), recursive = c(FALSE, TRUE)), tags = c("resolve", "lazy"), {
+make_test(title = "resolve()", args = list(lazy = c(FALSE, TRUE), result = c(FALSE, TRUE), recursive = c(FALSE, TRUE)), tags = c("resolve", "lazy"), {
   f <- future({
     Sys.sleep(0.5)
     list(a = 1, b = 42L)
   }, lazy = lazy)
-  res <- resolve(f, value = value, recursive = recursive)
+  res <- resolve(f, result = result, recursive = recursive)
   stopifnot(identical(res, f))
 })
 
 
-make_test(title = "resolve() - run-time exception", args = list(lazy = c(FALSE, TRUE), value = c(FALSE, TRUE), recursive = c(FALSE, TRUE)), tags = c("resolve", "lazy"), {
+make_test(title = "resolve() - run-time exception", args = list(lazy = c(FALSE, TRUE), result = c(FALSE, TRUE), recursive = c(FALSE, TRUE)), tags = c("resolve", "lazy"), {
   f <- future(list(a = 1, b = 42L, c = stop("Nah!")), lazy = lazy)
-  res <- resolve(f, value = value, recursive = recursive)
+  res <- resolve(f, result = result, recursive = recursive)
   stopifnot(identical(res, f))
 })